fix: number integer for weight direction messsage and toast added

+++ source/src/components/widgets/create/DirectionCreate.tsx
@@ -57,7 +57,7 @@ export const DirectionCreate = ({ onOpenChange }: { onOpenChange: (state: boolea
         merchant: z.string().min(1, translate("resources.direction.errors.merchant")),
         provider: z.string().min(1, translate("resources.direction.errors.provider")),
         terminal: z.string().min(1, translate("resources.direction.errors.terminal")),
-        weight: z.coerce.number()
+        weight: z.coerce.number().int()
     });
 
     const form = useForm<z.infer<typeof formSchema>>({
